Create best of two
Chef took an examination two times. In the first attempt, he scored 
𝑋
X marks while in the second attempt he scored 
𝑌
Y marks. According to the rules of the examination, the best score out of the two attempts will be considered as the final score.

Determine the final score of the Chef.